requirements.txt
openai-agents==0.2.6
pydantic==2.10.6
python-dotenv==1.0.0
rich==13.7.0
weasyprint==65.0
PyYAML==6.0.2